WIP: perf: use a dedicated OS thread to drop provider instances
